backend/requirements.txt
Flask
flask-cors
requests
google-genai
python-dotenv
